Many performance improvements
This is the result of several days of looking at traces and completely
redesigning things (for the better, turtle?) to at least make it faster
for reasonably real-world use cases.

1. Compressing its own `.venv` dir, which is open/stat-bound
2. Compressing 4GB of /dev/urandom, which is cpu-bound

Both on btrfs ssd but with hot filesystem cache so the choice of
filesystem and disk shouldn't matter.

Primarily tested on a Xeon E3-1231 v3 (4 core, 8 thread) machine, and
also on an i5-8350U laptop (similarly 4 core, 8 thread).