Provide context window in deflate compression
This is a very small win for large files with repetitive data, but is
effectively free.  Only affects ratio on files > `THREAD_BLOCK_SIZE`
(default 1 MiB).

Trying on the `/usr/lib/libx26[45].so` on my system (~24 MiB), I see a
0.08% ratio improvement.  If you have text files larger than
`THREAD_BLOCK_SIZE` you will probably notice a bigger difference.  This
unlocks being able to get closer to ideal ratio for small files by
turning `THREAD_BLOCK_SIZE` down as low as 32 KiB, which was giving up a
lot of ratio before.

(Actually lowering it that much appears to still be blocked on Python
locks including the GIL.)